Remove lazy initialization
Using Koin DI to inject `SplitTunneling` leads to some issues. By
default, injecting is lazy evaluated, which means that it was only being
initialized when trying to use `SplitTunneling`, which led to a deadlock
issue. Another issue with the lazy evaluation is that the
`ServiceConnnection` class loses control of the initialization order,
and can also miss the initial events sent from the service for
`SplitTunneling`.